requirements.txt
streamlit>=1.44.0,<2.0.0
pandas>=2.2.3,<3.0.0
scikit-learn>=1.6.1,<2.0.0
scipy>=1.15.2,<2.0.0
nltk>=3.9.1,<4.0.0
numpy>=1.26.4,<2.0.0
seaborn>=0.13.2,<0.14.0
statsmodels>=0.14.4,<0.15.0
transformers>=4.50.3,<5.0.0
sentence-transformers>=4.0.1,<5.0.0
fastapi>=0.115.12,<0.116.0
pdfplumber>=0.11.6,<0.12.0
plotly>=6.0.1,<7.0.0
geopandas>=1.0.1,<2.0.0
folium>=0.19.5,<0.20.0
dash>=3.0.1,<4.0.0
keplergl>=0.3.7,<0.4.0
shap>=0.47.1,<0.48.0
lime>=0.2.0.1,<0.3.0.0
xgboost>=3.0.0,<4.0.0
lightgbm>=4.6.0,<5.0.0
h3>=4.2.2,<5.0.0
streamlit-folium>=0.24.0,<0.25.0
streamlit-plotly-events>=0.0.6,<0.0.7
streamlit-lottie>=0.0.5,<0.0.6
dask>=2025.3.0,<2026.0.0
ray>=2.44.1,<3.0.0
catboost>=1.2.7,<2.0.0
shapely>=2.0.7,<3.0.0
cartopy>=0.24.1,<0.25.0
geoplot>=0.5.1,<0.6.0
optuna>=4.2.1,<5.0.0
python-dotenv>=1.1.0,<2.0.0
google-genai>=1.8.0,<2.0.0
pymongo>=4.11.3,<5.0.0
loguru<=0.7.3
mlflow>=2.21.2,<3.0.0
branca>=0.8.1,<0.9.0